move dual_stats_manager to solve compat issues
