+++ providers/edge3/src/airflow/providers/edge3/models/edge_worker.py
@@ -26,7 +26,7 @@ from typing import TYPE_CHECKING
 from sqlalchemy import Integer, String, delete, select
 from sqlalchemy.orm import Mapped
 
-from airflow.dual_stats_manager import DualStatsManager
+from airflow.metrics.dual_stats_manager import DualStatsManager
 from airflow.models.base import Base
 from airflow.providers.common.compat.sdk import AirflowException, Stats, timezone
 from airflow.providers.common.compat.sqlalchemy.orm import mapped_column
